Add HTTPS redirection to CloudFront distribution
An import statement for ViewerProtocolPolicy has been added in the cloudfront-deploy-stack.ts file. Furthermore, the configuration of the CloudFront distribution has been modified to add a viewer protocol policy that redirects all viewer requests to HTTPS. This change ensures that the communication between the viewers and the distribution is always encrypted, providing better security.