Update AWS CDK dependencies and refactor CloudFront origin handling
This commit updates the AWS CDK library versions in 'package.json' and 'package-lock.json' to the latest releases. Additionally, it refactors the CloudFront origin handling in 'cloudfront-deploy-stack.ts' to utilize the new S3BucketOrigin with Origin Access Control (OAC), enhancing security and simplifying the configuration. This change replaces the previous use of Origin Access Identity (OAI) for better integration with S3.